for #2084, add SimpleExpressionSegment
