+++ sharding-core/sharding-core-parse/sharding-core-parse-common/src/main/java/org/apache/shardingsphere/core/parse/antlr/sql/segment/dml/expr/LiteralExpressionSegment.java
@@ -34,7 +34,7 @@ import org.apache.shardingsphere.core.parse.old.parser.expression.SQLTextExpress
  */
 @RequiredArgsConstructor
 @Getter
-public final class LiteralExpressionSegment implements ExpressionSegment {
+public final class LiteralExpressionSegment implements SimpleExpressionSegment {
     
     private final int startIndex;
