+++ sharding-core/sharding-core-parse/sharding-core-parse-common/src/main/java/org/apache/shardingsphere/core/parse/antlr/sql/segment/dml/expr/ParameterMarkerExpressionSegment.java
@@ -29,7 +29,7 @@ import org.apache.shardingsphere.core.parse.old.parser.expression.SQLPlaceholder
  */
 @RequiredArgsConstructor
 @Getter
-public final class ParameterMarkerExpressionSegment implements ExpressionSegment {
+public final class ParameterMarkerExpressionSegment implements SimpleExpressionSegment {
     
     private final int stopIndex;
